Bulk dependency updates (PRs 46-50+) and fix import issues after restructure
