+++ tests/dev/test_resource_system.py
@@ -7,10 +7,10 @@ import sys
 from pathlib import Path
 
 # Add src to path for imports
-sys.path.insert(0, str(Path(__file__).parent / "src"))
+sys.path.insert(0, str(Path(__file__).parent.parent.parent))
 
-from deterministic_id_generator import DeterministicIDGenerator
-from resource_manager import ResourceRegistry, CacheManager, ResourceRecovery
+from src.deterministic_id_generator import DeterministicIDGenerator
+from src.resource_manager import ResourceRegistry, CacheManager, ResourceRecovery
 
 
 def test_resource_system():
