+++ tests/test_async_interface.py
@@ -14,9 +14,9 @@ from pathlib import Path
 from unittest.mock import Mock, patch, AsyncMock
 
 # Add src to path
-sys.path.insert(0, str(Path(__file__).parent.parent / "src"))
+sys.path.insert(0, str(Path(__file__).parent.parent))
 
-from async_interface import (
+from src.async_interface import (
     AsyncMCPInterface,
     AsyncTaskQueue,
     AsyncTask,
