+++ tests/test_haiku_integration.py
@@ -11,9 +11,9 @@ import sys
 from pathlib import Path
 
 # Add src to path
-sys.path.insert(0, str(Path(__file__).parent / "src"))
+sys.path.insert(0, str(Path(__file__).parent.parent))
 
-from haiku_subagent import (
+from src.haiku_subagent import (
     HaikuSubagent, ProcessingStrategy, CostLimits, VideoAnalysis
 )
 
@@ -134,13 +134,13 @@ def test_imports():
     
     try:
         # Test importing server modifications
-        from server import haiku_agent
+        from src.server import haiku_agent
         print("✅ Server integration import works")
     except ImportError as e:
         print(f"⚠️  Server import issue (expected if not running): {e}")
     
     # Test core imports
-    from haiku_subagent import yolo_smart_concat
+    from src.haiku_subagent import yolo_smart_concat
     print("✅ Core function imports work")
     
     try:
