+++ tests/test_hanging_issues.py
@@ -12,7 +12,7 @@ import sys
 import os
 
 # Add src to path for imports
-sys.path.insert(0, 'src')
+sys.path.insert(0, str(Path(__file__).parent.parent))
 
 logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
 logger = logging.getLogger(__name__)
